backend/requirements.txt
chromadb
sentence-transformers
langchain
langchain-openai
langchain-chroma
langgraph
fastapi
uvicorn[standard]
python-dotenv
